pyproject.toml
[tool.poetry]
name = "bank-queue"
version = "0.1.0"
description = "Web API of banking queue"
authors = ["Bekhan bekhanrayamjanov@gmail.com"]
readme = "README.md"
packages = [{include = "bank_queue"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "1.20.0"
django-filter = "2.4.0"
django-cors-headers = "3.8.0"
drf-spectacular = "^0.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
